refactor: 로그아웃 action 함수 분리
