+++ components/organism/MyPage/SignOutButton/SignOutButton.tsx
@@ -1,22 +1,5 @@
 import React from 'react';
-import { redirect } from 'next/navigation';
-import { cookies } from 'next/headers';
-
-export const signOutUser = async () => {
-  'use server';
-
-  const res = await fetch(`${process.env.NEXT_PUBLIC_BASE_URL}/auth/logout/`, {
-    method: 'POST',
-  });
-
-  if (!res.ok) {
-    throw new Error('Failed to fetch data');
-  }
-
-  cookies().delete('access');
-  cookies().delete('refresh');
-  redirect('/signin');
-};
+import { signOutUser } from '@/app/_actions/signOutUser';
 
 const SignOutButton = () => {
   return (
